requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
email-validator
python-dotenv 